[YDB#61] Fix relink/shmalloc subtest failure on ARM due to 8-byte shmsize (accompanying RELINKCTLERR error) having leading 4-byte garbage; This is because size_t is 4-byte on ARM whereas sprintf is done for %llx assuming it is 8-byte; Typecast input to (unsigned long long); As part of this, examine all %ll usages in codebase and fix them as needed

+++ sr_port/mur_merge_sort_extfmt.c
@@ -364,8 +364,8 @@ int mur_merge_sort_extfmt(void)
 				if (-1 == (sm_long_t)shm_ptr)
 				{
 					save_errno = errno;
-					SNPRINTF(errstr, SIZEOF(errstr),
-						"shmat() : shmid=%d shmsize=0x%llx", shmid, shm_rctl->jnlext_shm_size);
+					SNPRINTF(errstr, SIZEOF(errstr), "shmat() : shmid=%d shmsize=0x%llx",
+								shmid, (unsigned long long)shm_rctl->jnlext_shm_size);
 					gtm_putmsg_csa(CSA_ARG(NULL) VARLSTCNT(8)
 								ERR_SYSCALL, 5, LEN_AND_STR(errstr), CALLFROM, save_errno);
 					goto cleanup;
@@ -535,7 +535,7 @@ int mur_merge_sort_extfmt(void)
 					rctl = &rctl_start[index];
 					fn = ((fi_type *)rctl->file_info[recstat])->fn;
 					SNPRINTF(errstr, SIZEOF(errstr), "fread() : %s : Expected = %lld : Actual = %lld",
-													fn, jm_size, ret_size);
+										fn, (long long)jm_size, (long long)ret_size);
 					if (save_errno)
 					{	/* ERROR encountered during GTM_FREAD */
 						gtm_putmsg_csa(CSA_ARG(NULL) VARLSTCNT(8)
@@ -555,7 +555,7 @@ int mur_merge_sort_extfmt(void)
 					assert(FALSE);
 					assert(save_errno);
 					SNPRINTF(errstr, SIZEOF(errstr), "fwrite() : %s : Expected = %lld : Actual = %lld",
-												fn_out, jm_size, ret_size);
+										fn_out, (long long)jm_size, (long long)ret_size);
 					gtm_putmsg_csa(CSA_ARG(NULL) VARLSTCNT(8)
 								ERR_SYSCALL, 5, LEN_AND_STR(errstr), CALLFROM, save_errno);
 					goto cleanup;
